Fixed a `repeatCount` problem in 64bit architecture.

+++ PulsingHalo/PulsingHaloLayer.h
@@ -16,8 +16,8 @@
 @property (nonatomic, assign) CGFloat radius;                   // default: 60pt
 @property (nonatomic, assign) NSTimeInterval animationDuration; // default: 3s
 @property (nonatomic, assign) NSTimeInterval pulseInterval;     // default: 0s
-@property (nonatomic, assign) CGFloat repeatCount;              // default: INFINITY
+@property (nonatomic, assign) float repeatCount;                // default: INFINITY
 
-- (id)initWithRepeatCount:(CGFloat) repeatCount;
+- (id)initWithRepeatCount:(float)repeatCount;
 
 @end
